Massive update of changes [wip]: - site, static & menu modules are now in frontend - user, role, permission management works again - system-wide request cache - system variables & menu items now use cache - install process now 100% hook based - revamped form & controller API (thx ViewKit) - new get endpoints & view pages

+++ Sources/FeatherCore/Modules/User/Controllers/UserAdminController.swift
@@ -16,37 +16,29 @@ struct UserAdminController: ViperAdminViewController {
         Model.FieldKeys.email,
     ]
 
-    func findModelBy(id: Model.IDValue, req: Request) -> EventLoopFuture<Model> {
-        Model.findWithRolesBy(id: id, on: req.db).unwrap(or: Abort(.notFound))
+    func listQuery(search: String, queryBuilder: QueryBuilder<Model>, req: Request) {
+        queryBuilder.filter(\.$email ~~ search)
     }
 
-    func find(_ req: Request) throws -> EventLoopFuture<Model> {
-        guard
-            let id = req.parameters.get(idParamKey),
-            let uuid = UUID(uuidString: id)
-        else {
-            throw Abort(.badRequest)
-        }
-        return findModelBy(id: uuid, req: req)
+    func findBy(_ id: UUID, on db: Database) -> EventLoopFuture<Model> {
+        Model.findWithRolesBy(id: id, on: db).unwrap(or: Abort(.notFound, reason: "User not found"))
     }
 
-    func searchList(using qb: QueryBuilder<Model>, for searchTerm: String) {
-        qb.filter(\.$email ~~ searchTerm)
+    func afterCreate(req: Request, form: CreateForm, model: Model) -> EventLoopFuture<Model> {
+        findBy(model.id!, on: req.db)
     }
 
-    func beforeList(req: Request, queryBuilder: QueryBuilder<Model>) throws -> QueryBuilder<Model> {
-        queryBuilder.sort(\Model.$email)
-    }
-
-    func afterCreate(req: Request, form: UserEditForm, model: UserModel) -> EventLoopFuture<UserModel> {
-        findModelBy(id: model.id!, req: req)
-    }
-
-    func afterUpdate(req: Request, form: UserEditForm, model: UserModel) -> EventLoopFuture<UserModel> {
-        findModelBy(id: model.id!, req: req)
+    func afterUpdate(req: Request, form: UpdateForm, model: Model) -> EventLoopFuture<Model> {
+        findBy(model.id!, on: req.db)
     }
 
     func beforeDelete(req: Request, model: Model) -> EventLoopFuture<Model> {
-        UserTokenModel.query(on: req.db).filter(\.$user.$id == model.id!).delete().map { model }
+        UserModel.query(on: req.db).count().flatMap { count in
+            if count == 1 {
+                return req.eventLoop.future(error: Abort(.badRequest, reason: "You can't delete every user"))
+            }
+            return UserTokenModel.query(on: req.db).filter(\.$user.$id == model.id!).delete().map { model }
+        }
     }
+
 }
